Simplify interrupt setup code
Doesn't include PIC driver.
