+++ src/kernel/interrupts/isr.cpp
@@ -1,13 +1,14 @@
 #include "interrupts/isr.hpp"
 
 #include "drivers/interrupts/pic.hpp"
+#include "interrupts/idt.hpp"
 #include "interrupts/interrupts.hpp"
 #include "logging/logger.hpp"
 
 #define ISR(NAME, BODY)                          \
     extern "C" void isr_##NAME(void *reserved) { \
-        Interrupts::disable();                   \
-        BODY Interrupts::enable();               \
+        DISABLE_INTERRUPTS();                    \
+        BODY ENABLE_INTERRUPTS();                \
     }
 
 #define PIC_ISR(NAME, BODY, ID) \
@@ -15,7 +16,7 @@
 
 ISR(divide_by_zero, logging::error("Divide by zero error!");)
 
-PIC_ISR(acknowledge_interrupt, , Interrupt::PIC_TIMER);
+PIC_ISR(acknowledge_interrupt, , interrupts::Id::PIC_TIMER);
 
 PIC_ISR(keyboard_press, logging::info("Key was pressed!\n");
-        , Interrupt::PIC_KEYBOARD)
+        , interrupts::Id::PIC_KEYBOARD)
